backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dat9uy <46886826+dat9uy@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
pydantic = { extras = ["email-validator"], version = "^1.8.1" }
databases = { extras = ["postgresql"], version = "^0.4.2" }
loguru = "^0.5.3"
yoyo-migrations = "^7.3.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^20.8b1"
pytest-asyncio = "^0.14.0"
httpx = "^0.17.1"
asgi-lifespan = "^1.0.1"
testcontainers = { extras = ["postgresql"], version = "^3.3.0" }
docker = "^4.4.4"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
#filterwarnings = "error"
#addopts = '''
#  --strict-markers
#  --tb=short
#  --cov=app
#  --cov=tests
#  --cov-branch
#  --cov-report=term-missing
#  --cov-report=html
#  --cov-report=xml
#  --no-cov-on-fail
#  --cov-fail-under=100
#'''
env_files = ["./.test.env"]
